@@ -4,6 +4,7 @@
     "type": "git",
     "url": "https://github.com/ild-games/duckling"
   },
+  "type": "commonjs",
   "version": "0.0.1",
   "scripts": {
     "start": "npm run build && concurrently \"npm run watch\" \"npm run duckling\" ",
@@ -23,56 +24,59 @@
   },
   "license": "ISC",
   "dependencies": {
-    "@angular/animations": "5.2.10",
-    "@angular/cdk": "5.2.5",
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/forms": "5.2.10",
-    "@angular/http": "5.2.10",
-    "@angular/material": "5.2.5",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@angular/upgrade": "5.2.10",
-    "@types/glob": "5.0.35",
-    "@types/hammerjs": "2.0.35",
-    "@types/howler": "2.0.5",
-    "@types/lodash": "4.14.108",
-    "@types/mathjs": "3.20.0",
-    "@types/node": "8.0.53",
-    "@types/pixi.js": "4.7.3",
-    "@types/webfontloader": "1.6.29",
-    "es6-shim": "0.35.3",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "7.2.15",
+    "@angular/material": "8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/upgrade": "8.2.14",
+    "@pixi/canvas-renderer": "^5.2.0",
+    "es6-shim": "0.35.5",
     "font-awesome": "4.7.0",
-    "glob": "7.1.2",
-    "gulp": "3.9.1",
-    "gulp-sass": "4.0.1",
-    "gulp-watch": "5.0.0",
-    "howler": "2.0.9",
-    "immutable": "3.8.2",
-    "lodash": "4.17.10",
-    "mathjs": "4.1.2",
+    "glob": "7.1.6",
+    "gulp": "4.0.2",
+    "gulp-sass": "5.0.0",
+    "gulp-watch": "5.0.1",
+    "hammerjs": "^2.0.8",
+    "howler": "2.1.2",
+    "immutable": "4.0.0-rc.12",
+    "lodash": "4.17.15",
+    "mathjs": "6.2.5",
     "open-iconic": "1.1.1",
-    "pixi.js": "4.7.3",
-    "redux": "4.0.0",
-    "reflect-metadata": "0.1.12",
-    "rxjs": "5.5.0",
-    "systemjs": "0.21.3",
+    "pixi.js": "5.2.0",
+    "redux": "4.0.4",
+    "reflect-metadata": "0.1.13",
+    "rxjs": "6.5.3",
+    "rxjs-compat": "6.5.3",
+    "sass": "1.45.1",
+    "systemjs": "6.1.5",
     "webfontloader": "1.6.28",
-    "zone.js": "0.8.26"
+    "zone.js": "0.10.2"
   },
   "devDependencies": {
-    "@types/chai": "4.1.3",
-    "@types/chai-spies": "0.0.0",
-    "@types/mocha": "5.2.0",
-    "chai": "4.1.2",
+    "@types/chai": "4.2.5",
+    "@types/chai-spies": "1.0.1",
+    "@types/glob": "7.1.1",
+    "@types/hammerjs": "2.0.36",
+    "@types/howler": "2.1.2",
+    "@types/lodash": "4.14.149",
+    "@types/mathjs": "6.0.2",
+    "@types/mocha": "5.2.7",
+    "@types/node": "12.12.14",
+    "@types/webfontloader": "1.6.29",
+    "chai": "4.2.0",
     "chai-spies": "1.0.0",
-    "concurrently": "3.5.1",
-    "electron": "1.8.6",
-    "electron-mocha": "6.0.2",
-    "mocha": "5.1.1",
-    "prettier": "1.13.7",
-    "typescript": "2.8.3"
+    "concurrently": "5.0.0",
+    "electron": "11.0.0",
+    "electron-mocha": "8.1.2",
+    "mocha": "6.2.2",
+    "prettier": "1.19.1",
+    "typescript": "4.5.4"
   },
   "prettier": {
     "tabWidth": 4,
