@@ -1,9 +1,11 @@
+import 'hammerjs';
 import { NgModule, ApplicationRef, Type } from "@angular/core";
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 import { BrowserModule } from "@angular/platform-browser";
 import { FormsModule } from "@angular/forms";
-import { remote } from "electron";
+import { remote } from 'electron';
 import { SCALE_MODES } from "pixi.js";
+import * as PIXI from "pixi.js";
 
 import { EntityDrawerService } from "./canvas/drawing/entity-drawer.service";
 import { RenderPriorityService } from "./canvas/drawing/render-priority.service";
